server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.6.0
astroid==2.15.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.6
dashscope==1.11.0
decorator==5.1.1
dill==0.3.6
Django==3.2.18
django-cors-headers==4.1.0
django-jet-4==0.0.1
djangorestframework==3.14.0
frozenlist==1.3.3
google-api-python-client==1.4.1
httplib2==0.21.0
idna==3.4
importlib-metadata==6.5.0
ipython==7.34.0
isort==5.11.5
jedi==0.18.2
joblib==1.3.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
nltk==3.8.1
oauth2client==4.1.3
openai==0.27.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pylint==2.17.1
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytz==2022.7.1
regex==2023.8.8
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
thulac==0.2.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
traitlets==5.9.0
typed-ast==1.5.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
